+++ src/Database/PostgreSQL/PQTypes/Checks.hs
@@ -542,7 +542,12 @@ checkDBStructure options tables = fmap mconcat . forM tables $ \(table, version)
               ]
 
             (localIndexes, indexes) = (`partition` allIndexes) $ \(_, name) ->
-              "local_" `T.isPrefixOf` unRawSQL name
+              -- Manually created indexes for ad-hoc improvements.
+                 "local_" `T.isPrefixOf` unRawSQL name
+              -- Indexes related to the REINDEX operation, see
+              -- https://www.postgresql.org/docs/current/sql-reindex.html.
+              || "_ccnew" `T.isSuffixOf` unRawSQL name
+              || "_ccold" `T.isSuffixOf` unRawSQL name
 
         checkForeignKeys :: [ForeignKey] -> [(ForeignKey, RawSQL ())]
                          -> ValidationResult